Add files via upload
This page has been translated into Korean. I thought that Google's automatic translation might cause confusion by translating key words and unique words, so I used the original text as is.
We translated it a little more smoothly.
Next to the community move button, we also created a button to go to the translated page.